+++ outputs.tf
@@ -1,16 +1,21 @@
+output "cluster_arn" {
+  value       = aws_ecs_cluster.this.arn
+  description = "string ||| AWS Arn for the Fargate cluster."
+}
+
 output "cluster_name" {
   value       = aws_ecs_cluster.this.name
-  description = "string ||| "
+  description = "string ||| Name of the Fargate cluster."
 }
 
 output "service_discovery_id" {
   value       = aws_service_discovery_private_dns_namespace.service.id
-  description = "string ||| "
+  description = "string ||| AWS ID for the Private DNS namespace created for this cluster."
 }
 
 output "cluster_execution_role_name" {
   value       = aws_iam_role.execution.name
-  description = "string ||| "
+  description = "string ||| Name of AWS Role that is usually attached to each service for execution."
 }
 
 output "deployer" {