+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.3.0 (Apr 13, 2021)
 
 * Added `deployer` user with explicit permission to deploy to this fargate cluster.
+* Added `cluster_arn` to outputs.